switch to an asynchronous fetch engine
Fetch from memory into a FIFO, and then have the decoder
pull from the FIFO as needed with single-cycle latency.

This hides the latency from the memory pipelining required
to hit the timing requirements for the memory. It mostly
works, in that the device is much closer to running a sequence
flawlessly than ever before, but it still has occasional misfetches...