clocks: integrate Spencer's reconfigurable PLL
This imports Spencer's block for a PLL that can be
software reconfigured to switch between the 10 MHz PXI
chassis reference and the onboard 80 MHz oscillator.

It is not an exact import: it reworks a bunch of the input
register logic to remove a stage of unnecessary buffering
and to synchronously clear the command register rather
than asynchronously. (It is important that the command
lines be held high for one and only one cycle, so the
command register is set up to auto-clear itself the
cycle after it is written.)