output: separate the fan-out of the output
There was noise/crosstalk on the PXI lines.  This change is an attempt to
reduce that cross talk.  There is a possibility that we may need to yet
slow down the slew rate of the PXI pins specifically.  All output pins are
configured to with a drive strength of 8mA with no slew-control.  We have
two options if this patch does not work:
  - change the slew setting to slow down the PXI pins
  - change the output drive strength to 4mA (I prefer this because 8mA
    seems somewhat high and slightly more threatening for damaging PXI
    trigger pins if double-driven).